+++ docs-src/src/pages/tornado.tsx
@@ -16,121 +16,51 @@ import {
 import {
   EChartDirection,
   EGroupedBarLayout,
-  IAxes,
 } from '../../../src';
-import {
-  ITornadoData,
-  ITornadoProps,
-} from '../../../src/legacy/Tornado';
-import NativeTornado from '../../../src/Tornado';
-import { DeepPartial } from '../../../src/utils/types';
+import NativeTornado, { IProps } from '../../../src/Tornado';
+import { tornadoData } from '../../../test/fixtures';
 import JSXToString from '../components/JSXToString';
 import Layout from '../components/layout';
 import SEO from '../components/seo';
-import { grid } from '../data';
-
-const axis: DeepPartial<IAxes> = {
-  x: {
-    numberFormat: '.2s',
-    scale: 'linear',
-  },
-  y: {
-  },
-};
-
-const data2: ITornadoData = {
-  bins: ['Yes', 'No'],
-  // @Todo test with only one count set
-  counts: [
-    {
-      label: 'Background',
-      data: [
-        [0, 260], // Male bin 1, Male bin 2,
-        [0, 210], // Female bin 1, Female bin 2,
-      ],
-    },
-    {
-      label: 'Foreground',
-      data: [
-        [10, 2600], // Male bin 1, Male bin 2,
-        [330, 100], // Female bin 1, Female bin 2,
-      ],
-    },
-
-  ],
-}
 
 type Actions = { type: 'SET_SHOW_PERCENTAGES', show: boolean }
-  | { type: 'HERE' };
+  | { type: 'SET_DIRECTION', direction: EChartDirection };
 
-const initialState: DeepPartial<ITornadoProps> = {
-  data: {
-    bins: ['16-18', '18-25', '25-35', '35-50', '50-65', '65-∞'],
-    // @Todo test with only one count set
-    counts: [
-      {
-        label: 'Background',
-        data: [
-          [200, 2600, 5100, 9700, 8400, 6700], // Male bin 1, Male bin 2,
-          [2002, 2100, 4700, 8700, 4900, 1400], // Female bin 1, Female bin 2,
-        ],
-      },
-      {
-        label: 'Foreground',
-        data: [
-          [100, 260, 510, 970, 840, 670], // Male bin 1, Male bin 2,
-          [1000, 5500, 470, 870, 490, 140], // Female bin 1, Female bin 2,
-        ],
-      },
-
-    ],
-  },
-  axis,
+const initialState: IProps = {
+  data: tornadoData,
   splitBins: ['Male', 'Female'],
-  bar: {
-    grouped: {
-      paddingInner: 0.1,
-      paddingOuter: 0,
-    },
-    paddingInner: 0.1,
-    paddingOuter: 0,
-    overlayMargin: 0.2,
-    hover: {
-      lighten: 0.1,
-    },
-    // width: 40,
-  },
-  center: true,
-  delay: 0,
-  duration: 400,
-  grid,
   groupLayout: EGroupedBarLayout.OVERLAID,
-  width: '100%',
+  width: 600,
+  height: 500,
+  splitAxisHeight: 50,
+  xAxisHeight: 20,
+  direction: EChartDirection.HORIZONTAL,
+  showBinPercentages: false,
 }
 
-function reducer(state: ITornadoProps, action: Actions): ITornadoProps {
+function reducer(state: IProps, action: Actions): IProps {
   switch (action.type) {
     case 'SET_SHOW_PERCENTAGES':
       return {
         ...state,
         showBinPercentages: action.show,
       };
+    case 'SET_DIRECTION': {
+      return {
+        ...state,
+        direction: action.direction
+      }
+    }
   }
   return state;
 }
 
-
 const Tornado = () => {
-  const [state, dispatch] = useReducer(reducer, initialState as ITornadoProps);
+  const [state, dispatch] = useReducer(reducer, initialState as IProps);
   const [index, setIndex] = useState(0);
-  const [direction, setDIrection] = useState(EChartDirection.VERTICAL);
-  const chart = <NativeTornado {...state}
-    width={600}
-    direction={direction}
-    xAxisHeight={20}
-    showBinPercentages={state.showBinPercentages}
-    splitAxisHeight={50}
-    height={500} />
+  const chart = <NativeTornado
+    id="demo"
+    {...state} />
   return (
     <Layout>
       <SEO title="Line Chart" description="" />
@@ -170,14 +100,17 @@ const Tornado = () => {
             <FormControlLabel
               control={
                 <Switch
-                  checked={direction === EChartDirection.VERTICAL}
+                  checked={state.direction === EChartDirection.VERTICAL}
                   color="primary"
                   onChange={(_, value) => {
-                    setDIrection(direction === EChartDirection.VERTICAL ? EChartDirection.HORIZONTAL : EChartDirection.VERTICAL)
+                    dispatch({
+                      type: 'SET_DIRECTION',
+                      direction: state.direction === EChartDirection.VERTICAL ? EChartDirection.HORIZONTAL : EChartDirection.VERTICAL,
+                    })
                   }}
                 />
               }
-              label={`Direction ${direction}`}
+              label={`Direction ${state.direction}`}
             />
           </Grid>